+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Initial release
 - Update README.md to focus on end-user
+- Update Data Model and Function to handle long sql